Too many different impl, standardise, with file support
